+++ integration-tests/tests/api.rs
@@ -1,5 +1,6 @@
-use alloy::network::ReceiptResponse;
+use alloy::network::{ReceiptResponse, TxSigner};
 use alloy::providers::Provider;
+use std::time::Duration;
 use zksync_os_integration_tests::Tester;
 use zksync_os_integration_tests::assert_traits::ReceiptAssert;
 use zksync_os_integration_tests::contracts::EventEmitter;
@@ -57,6 +58,39 @@ async fn get_code() -> anyhow::Result<()> {
         before_block_code.is_empty(),
         "deployed bytecode is not empty before deploy block"
     );
+
+    Ok(())
+}
+
+#[test_log::test(tokio::test)]
+async fn get_transaction_count() -> anyhow::Result<()> {
+    // Test that the node takes pending mempool transactions into account for `eth_getTransactionCount`
+    // We set block time to 5 seconds to make sure that transaction spends >5 seconds in the mempool.
+    // This gives us time to check that the node returns the correct transaction count.
+    let tester = Tester::builder()
+        .block_time(Duration::from_secs(5))
+        .build()
+        .await?;
+    let alice = tester.l2_wallet.default_signer().address();
+    let l2_provider = &tester.l2_provider;
+
+    // No existing transactions yet at the start
+    assert_eq!(l2_provider.get_transaction_count(alice).await?, 0);
+
+    let deploy_pending_tx = EventEmitter::deploy_builder(l2_provider.clone())
+        .send()
+        .await?;
+    // Pending transaction count takes pending transaction into account, so it's 1
+    assert_eq!(l2_provider.get_transaction_count(alice).pending().await?, 1);
+    // Latest transaction count is still 0
+    assert_eq!(l2_provider.get_transaction_count(alice).latest().await?, 0);
+    // Omitting block id defaults to latest block
+    assert_eq!(l2_provider.get_transaction_count(alice).await?, 0);
+
+    // Wait for the transaction to be mined and check that the transaction count is 1 now
+    deploy_pending_tx.expect_successful_receipt().await?;
+    assert_eq!(l2_provider.get_transaction_count(alice).pending().await?, 1);
+    assert_eq!(l2_provider.get_transaction_count(alice).latest().await?, 1);
 
     Ok(())
 }