+++ lib/rpc/src/lib.rs
@@ -13,15 +13,18 @@ mod result;
 mod rpc_storage;
 pub use rpc_storage::{ReadRpcStorage, RpcStorage};
 mod debug_impl;
+mod monitoring_middleware;
 mod sandbox;
 mod tx_handler;
 mod types;
 mod zks_impl;
 
 use crate::debug_impl::DebugNamespace;
+use crate::eth_call_handler::EthCallHandler;
 use crate::eth_filter_impl::EthFilterNamespace;
 use crate::eth_impl::EthNamespace;
 use crate::eth_pubsub_impl::EthPubsubNamespace;
+use crate::monitoring_middleware::Monitoring;
 use crate::ots_impl::OtsNamespace;
 use crate::zks_impl::ZksNamespace;
 use alloy::primitives::Address;
@@ -29,6 +32,7 @@ use anyhow::Context;
 use hyper::Method;
 use jsonrpsee::RpcModule;
 use jsonrpsee::server::{ServerBuilder, ServerConfigBuilder};
+use jsonrpsee::ws_client::RpcServiceBuilder;
 use tower_http::cors::{Any, CorsLayer};
 use zksync_os_mempool::L2TransactionPool;
 use zksync_os_rpc_api::debug::DebugApiServer;
@@ -48,8 +52,15 @@ pub async fn run_jsonrpsee_server<RpcStorage: ReadRpcStorage, Mempool: L2Transac
     tracing::info!("Starting JSON-RPC server at {}", config.address);
 
     let mut rpc = RpcModule::new(());
+    let eth_call_handler = EthCallHandler::new(config.clone(), storage.clone(), chain_id);
     rpc.merge(
-        EthNamespace::new(config.clone(), storage.clone(), mempool.clone(), chain_id).into_rpc(),
+        EthNamespace::new(
+            storage.clone(),
+            mempool.clone(),
+            eth_call_handler.clone(),
+            chain_id,
+        )
+        .into_rpc(),
     )?;
     rpc.merge(
         EthFilterNamespace::new(config.clone(), storage.clone(), mempool.clone()).into_rpc(),
@@ -57,7 +68,7 @@ pub async fn run_jsonrpsee_server<RpcStorage: ReadRpcStorage, Mempool: L2Transac
     rpc.merge(EthPubsubNamespace::new(storage.clone(), mempool).into_rpc())?;
     rpc.merge(ZksNamespace::new(bridgehub_address, storage.clone()).into_rpc())?;
     rpc.merge(OtsNamespace::new(storage.clone()).into_rpc())?;
-    rpc.merge(DebugNamespace::new(storage).into_rpc())?;
+    rpc.merge(DebugNamespace::new(storage, eth_call_handler).into_rpc())?;
 
     // Add a CORS middleware for handling HTTP requests.
     // This middleware does affect the response, including appropriate
@@ -71,6 +82,10 @@ pub async fn run_jsonrpsee_server<RpcStorage: ReadRpcStorage, Mempool: L2Transac
         .allow_headers([hyper::header::CONTENT_TYPE]);
     let middleware = tower::ServiceBuilder::new().layer(cors);
 
+    let max_response_size_bytes = config.max_response_size_bytes();
+    let rpc_middleware = RpcServiceBuilder::new()
+        .layer_fn(move |service| Monitoring::new(service, max_response_size_bytes));
+
     let server_config = ServerConfigBuilder::default()
         .max_connections(config.max_connections)
         .max_request_body_size(config.max_request_size_bytes())
@@ -78,7 +93,8 @@ pub async fn run_jsonrpsee_server<RpcStorage: ReadRpcStorage, Mempool: L2Transac
         .build();
     let server_builder = ServerBuilder::default()
         .set_config(server_config)
-        .set_http_middleware(middleware);
+        .set_http_middleware(middleware)
+        .set_rpc_middleware(rpc_middleware);
 
     let server = server_builder
         .build(config.address)